+++ packages/flutterwave/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/language-flutterwave
 
+## 1.0.7
+
+### Patch Changes
+
+- c5f8728: Update undici dependency
+- Updated dependencies [c5f8728]
+  - @openfn/language-common@3.3.4
+
 ## 1.0.6 - 27 May 2026
 
 ### Patch Changes