+++ packages/formsg/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-formsg
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [c5f8728]
+  - @openfn/language-common@3.3.4
+
 ## 1.0.10 - 27 May 2026
 
 ### Patch Changes