+++ packages/monnify/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/language-monnify
 
+## 1.1.2
+
+### Patch Changes
+
+- c5f8728: Update undici dependency
+- Updated dependencies [c5f8728]
+  - @openfn/language-common@3.3.4
+
 ## 1.1.1 - 27 May 2026
 
 ### Patch Changes