+++ packages/msgraph/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openfn/language-msgraph
 
+## 0.9.2
+
+### Patch Changes
+
+- c5f8728: Update undici dependency
+- Updated dependencies [c5f8728]
+  - @openfn/language-common@3.3.4
+
 ## 0.9.1 - 27 May 2026
 
 ### Patch Changes
@@ -385,7 +393,7 @@
 ### Patch Changes
 
 - aad9549: Ensure that standard OAuth2 credentials with snake-cased
-  "access\_token" keys can be used for OAuth2-reliant adaptors
+  "access_token" keys can be used for OAuth2-reliant adaptors
 - Updated dependencies \[aad9549]
   - @openfn/language-common@1.10.0