+++ packages/opencrvs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/language-opencrvs
 
+## 1.2.2
+
+### Patch Changes
+
+- c5f8728: Update undici dependency
+- Updated dependencies [c5f8728]
+  - @openfn/language-common@3.3.4
+  - @openfn/language-fhir-4@0.5.5
+
 ## 1.2.1 - 27 May 2026
 
 ### Patch Changes